// packages/contracts-clients/dao-pre-propose-single/DaoPreProposeSingle.types.ts
export type Uint128 = string;
export type DepositToken = {
  token: {
    denom: UncheckedDenom;
  };
} | {
  voting_module_token: {};
};
export type UncheckedDenom = {
  native: string;
} | {
  cw20: string;
};
export type DepositRefundPolicy = "always" | "only_passed" | "never";
export interface InstantiateMsg {
  deposit_info?: UncheckedDepositInfo | null;
  extension: Empty;
  open_proposal_submission: boolean;
}
export interface UncheckedDepositInfo {
  amount: Uint128;
  denom: DepositToken;
  refund_policy: DepositRefundPolicy;
}
export interface Empty {
  [k: string]: unknown;
}